+++ lib/matplotex/figure/areal.ex
@@ -395,6 +395,19 @@ defmodule Matplotex.Figure.Areal do
     Algebra.transform_given_point(x, y, sx, sy, tx, ty)
   end
 
+
+  def do_transform(%Dataset{x: x, y: y, bottom: bottom} = dataset, xlim, ylim, width, height, transition) when is_list(bottom) do
+    y = [y | bottom]|> Nx.tensor() |> Nx.transpose()|> Nx.to_list()
+
+    transformed =
+      x
+      |> Enum.zip(y)
+      |> Enum.map(fn {x, y} ->
+        transform_with_bottom(x,y, xlim, ylim, width, height, transition)
+      end)
+
+    %Dataset{dataset | transformed: transformed}
+  end
   def do_transform(%Dataset{x: x, y: y} = dataset, xlim, ylim, width, height, transition) do
     transformed =
       x
@@ -407,4 +420,13 @@ defmodule Matplotex.Figure.Areal do
 
     %Dataset{dataset | transformed: transformed}
   end
+
+
+  defp transform_with_bottom(x, y, xlim, ylim, width, height, transition) when is_list(y) do
+    y_top = Enum.sum(y)
+    y_bottom = y|>tl()|>Enum.sum()
+    transformed  = transformation(x, y_top, xlim, ylim, width, height, transition) |> Algebra.flip_y_coordinate()
+    {_, transformed_y_bottom} = transformation(x, y_bottom, xlim, ylim, width, height,transition) |> Algebra.flip_y_coordinate()
+    {transformed, transformed_y_bottom}
+   end
 end
